modificacion links menu, organizacion para las ventanas, uso de router-dom
