+++ src/componentes/navBar.css
@@ -1,0 +1,3 @@
+.link{
+    text-decoration:none;
+}
