+++ src/vistas/Home.js
@@ -1,0 +1,8 @@
+export default function Home(){
+
+    return(
+        <div>
+            <h1>Home</h1>
+        </div>
+    )
+}
