+++ src/vistas/Preguntas.js
@@ -1,0 +1,8 @@
+export default function Preguntas(props) {
+
+    return(
+        <div>
+            <h1>Preguntas</h1>
+        </div>
+    )
+}
