+++ src/vistas/SaberMas.js
@@ -1,0 +1,8 @@
+export default function SaberMas(props) {
+
+    return(
+        <div>
+            <h1>SaberMas</h1>
+        </div>
+    )
+}
