+++ src/vistas/Testimonios.js
@@ -1,0 +1,8 @@
+export default function Testimonios(props) {
+
+    return(
+        <div>
+            <h1>Testimonios</h1>
+        </div>
+    )
+}
